requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues


azure-functions

pandas
transformers
stop-words
mysql-connector==2.2.9
sshtunnel==0.4.0
spacy
tiktoken
openai

